workloads/ecom/requirements.txt
numpy
pandas
torch
faiss-cpu
transformers
transformers4rec[pytorch,nvtabular]

